+++ marxs/analysis/analysis.py
@@ -7,27 +7,48 @@ from astropy.stats import sigma_clipped_stats
 
 from ..optics import FlatDetector
 
-def sigma_clipped_std(data, **kwargs):
+def sigma_clipped_std(photons, colname='det_x', **kwargs):
     '''Return stddev of sigma-clipped data.
 
     Parameters
     ----------
-    data : np.array
-        unbinned input data
+    photons : `astropy.table.Table`
+        Input photon list.
+    colname : string
+        Column name of the photon distribution to be minimized.
+        The default is set for detectors that look for a grating signal
+        (which is dispersed in ``det_y`` direction).
     kwargs : see `astropy.stats.sigma_clipped_stats`
         All keyword arguments are passed to `astropy.stats.sigma_clipped_stats`.
     '''
-    mean, median, std = sigma_clipped_stats(data, **kwargs)
+    mean, median, std = sigma_clipped_stats(photons[colname].data, **kwargs)
     return std
 
 
-def find_best_detector_position(photons, col='det_x',
+def mean_width_2d(photons):
+    '''Return average distance from center of det_x, det_y distribution.
+
+    Parameters
+    ----------
+    photons : `astropy.table.Table`
+        Input photon list.
+    '''
+    x = photons['det_x'].data
+    y = photons['det_y'].data
+    r = np.linalg.norm(np.stack([x - x.mean(), y - y.mean()]), axis=0)
+    return np.sum(r) / len(r)
+
+
+
+def find_best_detector_position(photons,
                                 objective_func=sigma_clipped_std,
+                                objective_func_args={'colname': 'det_x'},
                                 orientation=np.eye(3), **kwargs):
     '''Numerically find the position of best focus.
 
     This routine places detectors at different positions and
     calculates the width of the photon distribution for each position.
+    The detector is moved perpendicular to its plane.
     As a side effect, ``photons`` will be set to the intersection with
     the last try of the detector position.
 
@@ -35,12 +56,10 @@ def find_best_detector_position(photons, col='det_x',
     ----------
     photons : `astropy.table.Table`
         Input photon list.
-    col : string
-        Column name of the photon distribution to be minimized.
-        The default is set for detectors that look for a grating signal
-        (which is dispersed in ``det_y`` direction).
     objective_func : function
         Function that accepts a np.array as input and return the width.
+    object_func_args : dict
+        Dict of any other options for the `objective_func`.
     rotation : np.array of shape (3,3)
         Rotation matrix for the detector. By default the detector is parallel
         to the yz plane of the global coordinate system (see `pos4d`).
@@ -49,15 +68,17 @@ def find_best_detector_position(photons, col='det_x',
     Returns
     -------
     opt : OptimizeResult
-        see `scipy.optimize.minimize`
+        see `scipy.optimize.minimize_scalar`
 
     '''
     def width(x, photons):
-        mdet = FlatDetector(position=np.array([x, 0, 0]), orientation=orientation, zoom=1e5, pixsize=1.)
+        mdet = FlatDetector(position=np.dot(orientation, np.array([x, 0, 0])),
+                            orientation=orientation,
+                            zoom=1e5, pixsize=1.)
         photons = mdet(photons)
-        return objective_func(photons[col].data)
+        return objective_func(photons, **objective_func_args)
 
-    return scipy.optimize.minimize(width, 0, args=(photons,), options={'maxiter': 20, 'disp': True},
+    return scipy.optimize.minimize_scalar(width, args=(photons,), options={'maxiter': 20},
                                    **kwargs)